LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-basic-expected.txt


FAIL --pf  assert_equals: offsetLeft expected 10 but got 99999
FAIL --pf flip-block assert_equals: offsetLeft expected 10 but got 99999
FAIL --pf flip-inline assert_equals: offsetLeft expected 360 but got 99999
FAIL --pf flip-block flip-inline assert_equals: offsetLeft expected 360 but got 99999
FAIL --pf flip-inline flip-block assert_equals: offsetLeft expected 360 but got 99999
FAIL --pf flip-start assert_equals: offsetLeft expected 20 but got 99999
FAIL --pf flip-block flip-start flip-inline assert_equals: offsetLeft expected 20 but got 99999
FAIL --pf flip-inline flip-start flip-block assert_equals: offsetLeft expected 20 but got 99999
FAIL --pf flip-start flip-block assert_equals: offsetLeft expected 20 but got 99999
FAIL --pf flip-inline flip-start assert_equals: offsetLeft expected 20 but got 99999
FAIL --pf flip-start flip-inline assert_equals: offsetLeft expected 340 but got 99999
FAIL --pf flip-block flip-start assert_equals: offsetLeft expected 340 but got 99999
FAIL --pf flip-start flip-block flip-inline assert_equals: offsetLeft expected 340 but got 99999
FAIL --pf flip-start flip-inline flip-block assert_equals: offsetLeft expected 340 but got 99999
FAIL --pf flip-inline flip-block flip-start assert_equals: offsetLeft expected 340 but got 99999
FAIL --pf flip-block flip-inline flip-start assert_equals: offsetLeft expected 340 but got 99999
